+++ py_gql/execution/executor.py
@@ -48,7 +48,7 @@ from ..schema import (
 )
 from ..utilities import coerce_argument_values, directive_arguments
 from .default_resolver import default_resolver as _default_resolver
-from .tracer import NullTracer, Tracer
+from .instrumentation import Instrumentation
 from .wrappers import GroupedFields, ResolveInfo, ResponsePath
 
 Resolver = Callable[..., Any]
@@ -109,7 +109,7 @@ class Executor:
         "_errors",
         "_default_resolver",
         "_middlewares",
-        "_tracer",
+        "_instrumentation",
     )
 
     def __init__(
@@ -121,7 +121,7 @@ class Executor:
         context_value: Any,
         default_resolver: Optional[Resolver] = None,
         middlewares: Optional[Sequence[Callable[..., Any]]] = None,
-        tracer: Optional[Tracer] = None,
+        instrumentation: Optional[Instrumentation] = None,
         **_: Any
         # fmt: on
     ):
@@ -152,7 +152,7 @@ class Executor:
         )  # type: Dict[Tuple[Field, _ast.Field], Dict[str, Any]]
         self._resolver_cache = {}  # type: Dict[Resolver, Resolver]
         self._middlewares = middlewares or []
-        self._tracer = tracer or NullTracer()
+        self._instrumentation = instrumentation or Instrumentation()
 
     def add_error(
         self,
@@ -373,16 +373,18 @@ class Executor:
             nodes,
         )
 
+        on_field_end = self._instrumentation.on_field(
+            parent_value, self.context_value, info
+        )
+
         def fail(err):
             self.add_error(err, path, node)
-            self._tracer.on_field_end(info)
+            on_field_end()
             return None
 
         def complete(res):
-            self._tracer.on_field_end(info)
+            on_field_end()
             return self.complete_value(field_definition.type, nodes, path, res)
-
-        self._tracer.on_field_start(info)
 
         try:
             coerced_args = self.argument_values(field_definition, node)